# integrations/chatgpt-custom-gpt/openapi-gpt-action.yaml
openapi: 3.1.0
info:
  title: OrionBelt Semantic Layer
  version: 1.8.2
  description: >
    Compile YAML semantic models (OBML) as analytical SQL across 8 database dialects.
    This API runs in single-model mode with a pre-loaded semantic model.
    Use the discovery endpoints to explore available dimensions, measures, and metrics,
    then compile queries using business names instead of raw SQL.
servers:
  - url: https://YOUR-ORIONBELT-URL
    description: OrionBelt Semantic Layer API
paths:
...
  /v1/dimensions/{name}:
    get:
      operationId: getDimension
      summary: Get a dimension by name
      description: >
        Returns details for a single dimension including its data object,
        column reference, result type, time grain, and synonyms.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: The dimension name (e.g. "Country", "Order Date")
      responses:
        "200":
          description: Dimension details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DimensionDetail"
        "404":
          description: Dimension not found
...
  /v1/measures/{name}:
    get:
      operationId: getMeasure
      summary: Get a measure by name
      description: >
        Returns details for a single measure including its aggregation type,
        expression, source columns, and synonyms.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: The measure name (e.g. "Revenue", "Order Count")
      responses:
        "200":
          description: Measure details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeasureDetail"
        "404":
          description: Measure not found
...
  /v1/explain/{name}:
    get:
      operationId: explainLineage
      summary: Explain lineage of a dimension, measure, or metric
      description: >
        Shows the full lineage chain for any named artefact: which data objects,
        columns, joins, and expressions contribute to it. Useful for understanding
        how a business concept maps to physical tables.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Name of the dimension, measure, or metric to explain
      responses:
        "200":
          description: Lineage explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExplainResponse"
        "404":
          description: Artefact not found
  /v1/find:
    post:
      operationId: searchModel
      summary: Search model artefacts by name or synonym
      description: >
        Fuzzy-search across dimensions, measures, and metrics by name or synonym.
        Use this when the user asks about a concept and you need to find the matching
        artefact (e.g. "sales" might match a measure with synonym "sales").
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Search term (name or synonym)
                  example: revenue
                types:
                  type: array
                  items:
                    type: string
                    enum: [dimension, measure, metric]
                  description: Filter by artefact type (default is all types)
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        name:
                          type: string
                        match_field:
                          type: string
                        score:
                          type: number
...
  /v1/query/sql:
    post:
      operationId: compileQuery
      summary: Compile a semantic query to SQL
      description: >
        Compiles a query that selects dimensions and measures by their business names
        into optimized SQL for the specified dialect. The query uses business concepts,
        not table/column names. OrionBelt automatically resolves joins, picks the right
        planner (star schema or composite fact layer), and generates correct SQL.

        The response includes the compiled SQL, an explain plan showing the planner's
        reasoning, and SQL validation status.

        IMPORTANT: dimensions and measures must be referenced by their exact names as
        returned by listDimensions and listMeasures. Use searchModel to find names if unsure.
      x-openai-isConsequential: false
      parameters:
        - name: dialect
          in: query
          required: false
          schema:
            type: string
            enum: [bigquery, clickhouse, databricks, dremio, duckdb, mysql, postgres, snowflake]
            default: postgres
          description: Target SQL dialect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
            examples:
              simple:
                summary: Simple query
                value:
                  select:
                    dimensions:
                      - Country
                    measures:
                      - Revenue
                  limit: 100
              filtered:
                summary: Query with filter and ordering
                value:
                  select:
                    dimensions:
                      - Country
                      - Order Date
                    measures:
                      - Revenue
                      - Order Count
                  where:
                    - dimension: Country
                      operator: "="
                      value: "Germany"
                  order_by:
                    - field: Revenue
                      direction: desc
                  limit: 50
      responses:
        "200":
          description: Compiled SQL with explain plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryCompileResponse"
        "422":
          description: Query resolution error (unknown dimension/measure, fanout, etc.)
components:
  schemas:
...
    DimensionDetail:
      type: object
      properties:
        name:
          type: string
          description: Business name of the dimension
        data_object:
          type: string
          description: Source data object (table)
        column:
          type: string
          description: Source column name
        result_type:
          type: string
          description: Data type (string, int, float, date, datetime, boolean)
        time_grain:
          type: string
          nullable: true
          description: Time grain if temporal (day, week, month, quarter, year)
        format:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        synonyms:
          type: array
          items:
            type: string
    MeasureDetail:
      type: object
      properties:
        name:
          type: string
          description: Business name of the measure
        result_type:
          type: string
        aggregation:
          type: string
          description: Aggregation function (sum, count, count_distinct, avg, min, max, etc.)
        expression:
          type: string
          nullable: true
          description: "Expression using {[DataObject].[Column]} syntax"
        columns:
          type: array
          items:
            type: object
            properties:
              dataObject:
                type: string
              column:
                type: string
        distinct:
          type: boolean
        total:
          type: boolean
          description: Whether a grand total column is also computed
        format:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        synonyms:
          type: array
          items:
            type: string
...
    ExplainResponse:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        lineage:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              name:
                type: string
              detail:
                type: string
                nullable: true
...
    QueryRequest:
      type: object
      required:
        - select
      properties:
        select:
          type: object
          description: Dimensions and measures to select (by business name)
          properties:
            dimensions:
              type: array
              items:
                type: string
              description: "Dimension names (e.g. [\"Country\", \"Order Date\"])"
            measures:
              type: array
              items:
                type: string
              description: "Measure names (e.g. [\"Revenue\", \"Order Count\"])"
        where:
          type: array
          description: WHERE filters on dimensions
          items:
            type: object
            properties:
              dimension:
                type: string
                description: Dimension name to filter on
              operator:
                type: string
                enum: ["=", "!=", ">", "<", ">=", "<=", "like", "not like", "in", "not in", "between", "is null", "is not null"]
              value:
                description: Filter value (string, number, or array for in/between)
        having:
          type: array
          description: HAVING filters on measures
          items:
            type: object
            properties:
              measure:
                type: string
                description: Measure name to filter on
              operator:
                type: string
                enum: ["=", "!=", ">", "<", ">=", "<="]
              value:
                description: Filter value (number)
        order_by:
          type: array
          description: ORDER BY clauses
          items:
            type: object
            properties:
              field:
                type: string
                description: Dimension or measure name
              direction:
                type: string
                enum: [asc, desc]
                default: asc
        limit:
          type: integer
          description: Maximum number of rows to return
          example: 100
    QueryCompileResponse:
      type: object
      properties:
        sql:
          type: string
          description: The compiled SQL query
        dialect:
          type: string
          description: The target SQL dialect used
        resolved:
          type: object
          properties:
            fact_tables:
              type: array
              items:
                type: string
            dimensions:
              type: array
              items:
                type: string
            measures:
              type: array
              items:
                type: string
        warnings:
          type: array
          items:
            type: string
        sql_valid:
          type: boolean
          description: Whether the SQL passed post-generation syntax validation
        explain:
          type: object
          nullable: true
          description: Explain plan showing planner reasoning
          properties:
            planner:
              type: string
              description: "Planner used: star or cfl"
            planner_reason:
              type: string
            base_object:
              type: string
            base_object_reason:
              type: string
            joins:
              type: array
              items:
                type: object
                properties:
                  from_object:
                    type: string
                  to_object:
                    type: string
                  join_columns:
                    type: array
                    items:
                      type: string
                  reason:
                    type: string
            has_totals:
              type: boolean